requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alpaca-trade-api==3.2.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.3
box2d-py==2.3.5
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
dateparser==1.2.1
deprecation==2.1.0
elegantrl==0.3.6
exchange-calendars==4.2.8
Farama-Notifications==0.0.4
filelock==3.16.1
FinRL==0.3.7
flatbuffers==25.1.24
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.4.0
gcloud==0.18.3
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h5py==3.11.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
joblib==1.4.2
jwcrypto==1.5.6
keras==2.13.1
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
networkx==3.1
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pygame==2.1.0
pyluach==2.2.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2025.1
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
soupsieve==2.6
SQLAlchemy==1.4.54
stable_baselines3==2.4.1
stockstats==0.4.0
swig==4.3.0
sympy==1.13.3
ta-lib==0.4.35
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.2.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==1.26.20
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.6
wrapt==1.17.2
wrds==3.1.6
yarl==1.15.2
yfinance==0.2.52
zipp==3.20.2
